Make application of marker metadata more robust (#980)
Since the search engine algorithm applies the meta data of the shorter
url A also to all urls that contain url A it could be possible that
multiple marker meta data were set for one BLP entry. This resulted
in two markers in the map.

This change makes sure that meta data will only applied to the longer
url of an BLP entry if they are contained in each other, ignoring the
protocol of that URL pair. 

https://redmine.informationgrid.eu/issues/980