Make UVP BLP data import more robust
Add check for marker URL intersection in BLP import to prevent meta data
override during crawl process. The metadata of URL A will be overriding
the metadata of URL B if URLB starts with URL A.

Make sure the index is deleted if the property
'ingrid.delete.before.crawl' is set to true.

https://redmine.informationgrid.eu/issues/1093