Accept invalid SSL certificates in iPlug SE (#1368)
A new option 'ingrid.accept.invalid.ssl.certificates' is introduced to accept invalid SSL certificates (defaults to true). If true a dummy SSL factory will be used that accepts all SSL certificates.

This is necessary to be able to crawl URLs with self signed
certificates or certificates that are not yet in the java default
certstore.

(This has been a specific problem when crawling BLP URLs.)