Add support of umlaut urls/domains in iPlug SE (#111)
This enables support of "umlaut urls" and "umlaut domains". The
URLs are encoded when they are exported from the url management
for the nutch process.

This commit only targets the export of the URLs in the right format.
The nutch process still needs to be tested with the encoded URLs.