Use limit URLs as base for metadata transfer. (#985)
Now the metadata added to an URL entry in the maintenance GUI
will be based on the limit URLs instead of the start URLs. This
fixes the problem that the metadata were only propagated to
the URLs that start with the start URL pattern.

Now metadata is propagated to all URLs that match the limit
patterns of an entry.

CAUTION: Overlapping limit URLs will result in multiple metadata entries
for the matching URLs.