Disable transfer from Elasticsearch fields to PD. (#1749)
If index communication is tunneled via an iBus connection, no
Elasticsearch client is present. In this case no fields from
the Elasticsearch index mapping are transferred to the plugdescription. This is due to a missing method in iBus
communication for obtaining the Elasticsearch mapping. Maybe this should be fixed in the future.